データ受信するところまでできた

+++ AudioStreamer/StreamingPlayer.m
@@ -113,7 +113,7 @@ static void checkError(OSStatus err,const char *message){
         char property[5];
         *(UInt32 *)property = CFSwapInt32HostToBig(err);
         property[4] = '\0';
-        NSLog(@"%s = %-4.4s,%d",message, property,err);
+        NSLog(@"%s = %-4.4s,%d",message, property,(int)err);
         exit(1);
     }
 }
@@ -163,9 +163,6 @@ void outputCallback( void                 *inClientData,
             break;
         }
     }
-    
-   
-      
 }
 -(void)recvAudio:(NSData *)data{
